Setting up a better structure for geo-related tasks and ui
